+++ src/Mapping/Statement.php
@@ -52,12 +52,12 @@ class Statement
     public $authority;
 
     /**
-     * @var int
+     * @var \DateTime|null
      */
     public $created;
 
     /**
-     * @var int
+     * @var \DateTime|null
      */
     public $stored;
 
@@ -91,7 +91,7 @@ class Statement
         $statement->version = $model->getVersion();
 
         if (null !== $model->getCreated()) {
-            $statement->created = $model->getCreated()->getTimestamp();
+            $statement->created = $model->getCreated();
         }
 
         if (null !== $result = $model->getResult()) {
@@ -140,11 +140,11 @@ class Statement
         }
 
         if (null !== $this->created) {
-            $created = new \DateTime('@'.$this->created);
+            $created = $this->created;
         }
 
         if (null !== $this->stored) {
-            $stored = new \DateTime('@'.$this->stored);
+            $stored = $this->stored;
         }
 
         if (null !== $this->context) {